Implements the `enforcer` CLI
TL;DR
-----

Provides a simple CLI to check if the license is expired

Details
-------

Adds the command `enforcer` that will exit non-zero for an expired
license and zero if the license is valid.  The command doesn't take any
arguments and locates the SDK using the environment variable
`REPLICATED_SDK_ENDPOINT`. It uses exponential backoff to avoid DDOSing
the `repllicated.app` endpoint (or your CNAME thereof).