Enhances to run as a sidecar (#9)
TL;DR
-----

Updates `enforce` to work as init or sidecar

Details
-------

Modifies the `enforce` CLI to allow running license enforcement as either an
init container or a sidecar. When running as a sidecar, the license is checked
periodcally after the first valid license check. Events are emitted as before,
with an additional event emitted whenever a valid license has a change in
expiration date.